database/requirements.txt
opencv-python
boto3
pymysql
python-dotenv
openai